Try to use u32

+++ src/cl/prover.rs
@@ -1401,7 +1401,7 @@ impl ProofBuilder {
             )))?
             .value()
             .to_dec()?
-            .parse::<i32>()
+            .parse::<u64>()
             .map_err(|_| {
                 IndyCryptoError::InvalidStructure(format!(
                     "Value by key '{}' has invalid format",
